User jquery's pause(), instead of trigger pause
find the video element and if it exists pause it with pause(). The
use of trigger gives a false sense of pausing. It just sends out the
event, we will never know if the video will be paused. But with pause()
we will have an error in console, if something is wrong.